Compute the project build command before execution
